// src/state/types.ts
export interface NFTTokenInfo {
  tierType: number;
  claimedAmount: BigNumber;
  lastClaimedTime: number;
  isExpired: boolean;
}

export interface SerializedNftUserData {
  allowance: string
  stakedBalanceByTier: number[]
  isExpiredByTier: number[]
  totalClaimedAmountByTier: number[]
  pendingRewardsByTier: string[]
  balanceOfWallet: number
}

export interface DeserializedFarmUserData {
  allowance: BigNumber
  tokenBalance: BigNumber
  stakedBalance: BigNumber
  earnings: BigNumber
  proxy?: {
    allowance: BigNumber
    tokenBalance: BigNumber
    stakedBalance: BigNumber
    earnings: BigNumber
  }
}

export interface SerializedFarm extends SerializedFarmPublicData {
  userData?: SerializedFarmUserData
}

export interface SerializedNft {
  userData?: SerializedNftUserData
  soldOutByTier: BigNumber[]
}
// ...
export interface SerializedNftsState {
  data: SerializedNft  
  isInitialized: boolean
  isLoading: boolean
}